Added memory and started cpu layout

+++ src/emulator/timer.c
@@ -33,7 +33,7 @@ void timer_update() {
     uint64_t end_time = platform_time_ns();
     uint64_t delta_time = end_time - start_time;
     if(delta_time <= P_CPU_NS / 2) {
-        platform_sleep_ns((P_CPU_NS / 2) - (delta_time));
+        platform_sleep_ns((uint64_t)(P_CPU_NS / 2) - (delta_time));
     }
 
     start_time = platform_time_ns();
@@ -43,7 +43,7 @@ void timer_update() {
     end_time = platform_time_ns();
     delta_time = end_time - start_time;
     if(delta_time <= P_CPU_NS / 2) {
-        platform_sleep_ns((P_CPU_NS / 2) - (delta_time));
+        platform_sleep_ns((uint64_t)(P_CPU_NS / 2) - (delta_time));
     }
 
 }
